// apps/client/src/types/types.ts
export type ContainerSize = Database["public"]["Enums"]["container_size"];
// ...
export interface Order {
  fullName: string;
  itemName: string;
  container: ContainerSize;
  weight: number;
  flavor: string;
  flavorLabel: string;
  protein: string;
  proteinLabel: string;
  quantity: number;
}